+++ servant/test/Servant/Utils/LinksSpec.hs
@@ -11,14 +11,10 @@ import Data.Proxy ( Proxy(..) )
 import Servant.API
 
 type TestApi =
-  -- Capture and query/matrix params
+  -- Capture and query params
        "hello" :> Capture "name" String :> QueryParam "capital" Bool :> Delete '[JSON] ()
 
-  :<|> "parent" :> MatrixParams "name" String :> "child"
-                :> MatrixParam "gender" String :> Get '[JSON] String
-
   -- Flags
-  :<|> "ducks" :> MatrixFlag "yellow" :> MatrixFlag "loud" :> Delete '[JSON] ()
   :<|> "balls" :> QueryFlag "bouncy" :> QueryFlag "fast" :> Delete '[JSON] ()
 
   -- All of the verbs
@@ -34,7 +30,6 @@ type TestLink3 = "parent" :> "child" :> Get '[JSON] String
 
 type BadTestLink = "hallo" :> "hi" :> Get '[JSON] Bool
 type BadTestLink2 = "greet" :> Get '[PlainText] Bool
-type BadTestLink3 = "parent" :> "child" :> MatrixFlag "male" :> Get '[JSON] String
 
 type BadTestLink' = "hello" :> "hi" :> Get '[OctetStream] Bool
 type BadTestLink'2 = "greet" :> Get '[OctetStream] Bool
@@ -54,7 +49,7 @@ shouldBeURI link expected =
 
 spec :: Spec
 spec = describe "Servant.Utils.Links" $ do
-    it "Generates correct links for capture query and matrix params" $ do
+    it "Generates correct links for capture query params" $ do
         let l1 = Proxy :: Proxy ("hello" :> Capture "name" String :> Delete '[JSON] ())
         apiLink l1 "hi" `shouldBeURI` "hello/hi"
 
@@ -63,24 +58,12 @@ spec = describe "Servant.Utils.Links" $ do
                                          :> Delete '[JSON] ())
         apiLink l2 "bye" (Just True) `shouldBeURI` "hello/bye?capital=true"
 
-        let l3 = Proxy :: Proxy ("parent" :> MatrixParams "name" String
-                                          :> "child"
-                                          :> MatrixParam "gender" String
-                                          :> Get '[JSON] String)
-        apiLink l3 ["Hubert?x=;&", "Cumberdale"] (Just "Edward?")
-            `shouldBeURI` "parent;name[]=Hubert%3Fx%3D%3B%26;\
-                           \name[]=Cumberdale/child;gender=Edward%3F"
 
-    it "Generates correct links for query and matrix flags" $ do
+    it "Generates correct links for query flags" $ do
         let l1 = Proxy :: Proxy ("balls" :> QueryFlag "bouncy"
                                          :> QueryFlag "fast" :> Delete '[JSON] ())
         apiLink l1 True True `shouldBeURI` "balls?bouncy&fast"
         apiLink l1 False True `shouldBeURI` "balls?fast"
-
-        let l2 = Proxy :: Proxy ("ducks" :> MatrixFlag "yellow"
-                                         :> MatrixFlag "loud" :> Delete '[JSON] ())
-        apiLink l2 True True `shouldBeURI` "ducks;yellow;loud"
-        apiLink l2 False True `shouldBeURI` "ducks;loud"
 
     it "Generates correct links for all of the verbs" $ do
         apiLink (Proxy :: Proxy ("get" :> Get '[JSON] ())) `shouldBeURI` "get"