+++ servant-client/CHANGELOG.md
@@ -7,6 +7,7 @@ HEAD
 * `client` now takes an explicit `Manager` argument.
 * Use `http-api-data` instead of `Servant.Common.Text`
 * Client functions now consider any 2xx succesful.
+* Remove matrix params.
 
 0.4.1
 -----